address feedback and add tests
