+++ src/dotnetTemplate/Microsoft.PowerShell.Standard.Module.Template/Microsoft.PowerShell.Standard.Module.Template/.template.config/template.json
@@ -1,17 +1,17 @@
 {
   "$schema": "http://json.schemastore.org/template",
   "author": "Microsoft Corporation",
-  "classifications": [ "Common", "Console", "PowerShell", "Module" ],
+  "classifications": [ "Library", "PowerShell", "Module" ],
   "name": "PowerShell Standard Module",
-  "identity": "Microsoft.PowerShell.PowerShellStandard.Template",
-  "groupIdentity": "Microsoft.PowerShell.PowerShellStandard.Template",
+  "identity": "Microsoft.PowerShell.Standard.Module.Template",
+  "groupIdentity": "Microsoft.PowerShell.Standard.Module.Template",
   "shortName": "psmodule",
   "tags": {
     "language": "C#",
     "type":"project"
   },
   "preferNameDirectory": true,
-  "sourceName": "Microsoft.PowerShell.PowerShellStandard.Template",
+  "sourceName": "Microsoft.PowerShell.Standard.Module.Template",
   "symbols":{
     "PowerShellStandardVersion":{
       "type": "parameter",
@@ -48,6 +48,6 @@
     }
   ],
   "primaryOutputs": [
-    { "path": "Microsoft.PowerShell.PowerShellStandard.Template.csproj" }
+    { "path": "Microsoft.PowerShell.Standard.Module.Template.csproj" }
   ]
 }
